Add local console channel
Added a local console channel.
events sent back to irc bot are in the form of
":admin!~admin@localhost PRIVMSG :message\r\n"

Added a mechanism to only start console and join channels when
identified to Nickserv if we use a registered nickname.
This is mandatory because if we join before identifying, Chanserv
will remove any flag previously set.

Furthermore, if we don't wait for nickserv to asking us the password and
directly start the console, we won't be able to ask the password
through the console because we'll be waiting for user input.