util.log: Open at appending instead of init
Log line were appended to a file opened at startup and never closed,
thus moving log file for rotation (ie logrotate) had for side effect
that the 'old' log continued to be written while the 'new' one was left
empty (and so making the log rotation utterly useless unless we manually
restart the bot)

Now it opens the log in append mode (to prevent the log from being wiped)
and add our log line, then we close it. It adds a little overhead due to
continuous opening-closing, but log appends are a quite rare event on
the machine scale so it is not a problem.